BXC-5726 update logic
